services/integration-hub/requirements.txt
# AutoGraph v3 - Integration Hub Dependencies

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Validation
pydantic==2.10.0
pydantic-settings==2.6.0

# HTTP Client
httpx==0.27.0

# Utilities
python-dotenv==1.0.1

# Security - Secrets Management
cryptography==44.0.0
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9
